+++ src/models/dbDoc.model.ts
@@ -1,0 +1,3 @@
+import type { DBDoc as DBDocImport } from '../stores/databaseV2/types/dbDoc'
+
+export type DBDoc = DBDocImport
