+++ src/models/dbEndpoints.ts
@@ -1,0 +1,5 @@
+// re-imports and re-exports
+import type { DBEndpoint } from '../stores/databaseV2/endpoints'
+
+export { DB_ENDPOINTS } from '../stores/databaseV2/endpoints'
+export type IDBEndpoint = DBEndpoint
