+++ src/models/discussion.models.tsx
@@ -1,4 +1,4 @@
-import type { DBDoc } from '../stores/databaseV2/types'
+import type { DBDoc } from './dbDoc.model'
 import type { IQuestion } from './question.models'
 
 export type IComment = {
