+++ src/models/question.models.tsx
@@ -1,8 +1,10 @@
 import type { IUploadedFileMeta } from '../stores/storage'
 import type { IConvertedFileMeta } from '../types'
-import type { DBDoc, IModerable, ISharedFeatures } from './common.models'
+import type { DBDoc } from './dbDoc.model'
+import type { IModerable } from './moderation.model'
 import type { IQuestionCategory } from './questionCategories.model'
 import type { ISelectedTags } from './tags.model'
+import type { ISharedFeatures } from './voteUseful.model'
 
 /**
  * Question retrieved from the database also include metadata such as _id, _created and _modified
