+++ src/models/userPreciousPlastic.models.tsx
@@ -1,5 +1,5 @@
 import type { ProfileTypeLabel } from '../modules/profile/types'
-import type { DBDoc } from './common.models'
+import type { DBDoc } from './dbDoc.model'
 import type { IUser } from './user.models'
 
 export type PlasticTypeLabel =
