+++ src/pages/Maps/map.service.test.ts
@@ -1,5 +1,5 @@
 import { faker } from '@faker-js/faker'
-import { DB_ENDPOINTS } from 'src/models'
+import { DB_ENDPOINTS } from 'src/models/dbEndpoints'
 import { describe, expect, it, vi } from 'vitest'
 
 import { mapPinService } from './map.service'
