+++ src/pages/Question/question.service.ts
@@ -10,8 +10,8 @@ import {
   where,
 } from 'firebase/firestore'
 import { IModerationStatus } from 'oa-shared'
+import { DB_ENDPOINTS } from 'src/models/dbEndpoints'
 
-import { DB_ENDPOINTS } from '../../models'
 import { firestore } from '../../utils/firebase'
 
 import type {
