+++ src/stores/common/module.store.ts
@@ -6,7 +6,8 @@ import { formatLowerNoSpecial, randomID } from 'src/utils/helpers'
 
 import { Storage } from '../storage'
 
-import type { IDBEndpoint, ILocation } from 'src/models/common.models'
+import type { ILocation } from 'src/models/common.models'
+import type { IDBEndpoint } from 'src/models/dbEndpoints'
 import type { ISelectedTags } from 'src/models/tags.model'
 import type { IConvertedFileMeta } from '../../types'
 import type { IRootStore } from '../RootStore'
