+++ src/stores/databaseV2/utils/getQueryOptions.ts
@@ -1,4 +1,4 @@
-import type { DBQueryOptions } from '../types'
+import type { DBQueryOptions } from '../types/dbDoc'
 
 /**
  * For mapping queries it is easiest to provide a common subset of defaults
