--- src/test/components/HowtoProvider.tsx
@@ -1,49 +1,0 @@
-import { Form } from 'react-final-form'
-import arrayMutators from 'final-form-arrays'
-import { Provider } from 'mobx-react'
-import { useCommonStores } from 'src/common/hooks/useCommonStores'
-import { FactoryHowto } from 'src/test/factories/Howto'
-import { vi } from 'vitest'
-
-vi.mock('src/common/hooks/useCommonStores', () => {
-  return {
-    useCommonStores: () => ({
-      stores: {
-        howtoStore: {
-          uploadStatus: {
-            Start: false,
-            Cover: false,
-            'Step Images': false,
-            Files: false,
-            Database: false,
-            Complete: false,
-          },
-          validateTitleForSlug: vi.fn(),
-          uploadHowTo: vi.fn(),
-        },
-        tagsStore: {
-          allTags: [
-            {
-              label: 'test tag 1',
-              image: 'test img',
-            },
-          ],
-        },
-      },
-    }),
-  }
-})
-
-export const HowtoProvider = ({ children }) => {
-  const formProps = {
-    formValues: FactoryHowto(),
-    onSubmit: vi.fn(),
-    mutators: { ...arrayMutators },
-    component: () => children,
-  }
-  return (
-    <Provider {...useCommonStores().stores}>
-      <Form {...formProps} />
-    </Provider>
-  )
-}
