--- src/test/components/index.ts
@@ -1,3 +1,0 @@
-export { HowtoProvider } from './HowtoProvider'
-export { ResearchProvider } from './ResearchProvider'
-export { SettingsProvider } from './SettingsProvider'
