+++ src/test/factories/Howto.ts
@@ -1,7 +1,7 @@
 import { faker } from '@faker-js/faker'
 import { DifficultyLevel, IModerationStatus } from 'oa-shared'
 
-import type { IHowtoDB, IHowtoStep } from 'src/models'
+import type { IHowtoDB, IHowtoStep } from 'src/models/howto.models'
 
 export const FactoryHowto = (
   howtoOverloads: Partial<IHowtoDB> = {},
